add equation processing (#16) Probably breaks the bot
